// apps/desktop/src/types/generated/CatalogChangelogEntryDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CatalogChangeDto } from "./CatalogChangeDto";

export type CatalogChangelogEntryDto = {
  id: string;
  title: string;
  version: string | null;
  text: string | null;
  changes: Array<CatalogChangeDto>;
  publishedAt: number;
};